Add tests and improve validation method
Finally figured out how to calculate if an image is optimal resolution or not. This method takes into account the actual resolution of the device screen. The trick is understanding that CSS calculates sizes based on 96ppi, independent of the device resolution.

Also added some files I can use to test.